+++ README.md
@@ -1,1 +1,13 @@
 # BASIC NODEJS API
+This is a very basic API using NodeJS, ExpressJS and MongoDB.
+
+The application runs using two containers: one for the database and other for the app itself.
+
+## Usage
+to start containers, run:
+''' 
+docker compose up -d
+''' 
+
+## Tests
+TBD
