+++ app/DbManager.js
@@ -1,18 +1,21 @@
 class DbManager{
 
     constructor(mongoose) {
+        this.Schema = mongoose.Schema;
+        // const dburl = 'mongodb://localhost:27017/toys';
+        const dburl = 'mongodb://db:27017/toys';
+        
+        try {
+            mongoose.connect(dburl);
+        } catch (error) {
+            handleError(error);
+        }
 
-        this.mongoose = mongoose;
-        this.Schema = this.mongoose.Schema;
-        this.dburl = 'mongodb://localhost:27017/toys';
+        return mongoose;
+    }
 
-        this.mongoose.connect(this.dburl,{ useNewUrlParser: true }, (err)=>{
-            console.log('Db connected. Go ahead.! ', err)
-        }).catch(function(err){
-            console.log(err)
-        })
-
-        return this.mongoose;
+    handleError(error){
+        console.log('Connection error: ', error);
     }
 }
