+++ app/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "node_sample_with_docker",
+  "name": "nodejs_express_basic_app_sample_with_docker",
   "version": "1.0.0",
   "description": "Node.js on Docker",
-  "author": "First Last <first.last@example.com>",
+  "author": "Gabriel Guerra <gabrielguerra@gmail.com>",
   "main": "server.js",
   "scripts": {
     "start": "node server.js"
@@ -10,7 +10,7 @@
   "dependencies": {
     "body-parser": "^1.19.0",
     "express": "^4.16.1",
-    "mongoose": "^5.9.7"
+    "mongoose": "^8.1.0"
   },
   "devDependencies": {
     "nodemon": "^2.0.3"
