+++ docker-compose.yml
@@ -4,8 +4,7 @@ services:
   node-api:
     container_name: node-api
     build: ./
-#    command: nodemon server.js
-    image: node:13
+    image: node:18
     ports:
       - "8080:8080"
       - "443:443"
@@ -22,5 +21,3 @@ services:
       - "27017"
     volumes:
       - ./data:/mongodata
-
-
