--- package.json
@@ -1,6 +1,0 @@
-{
-  "name": "basic-node-api",
-  "lockfileVersion": 3,
-  "requires": true,
-  "packages": {}
-}
